services/ml/production/qualite.py
"""
Module pour la prédiction de la qualité des récoltes.
"""

from typing import Dict, Any, List
from datetime import date
import numpy as np
from sqlalchemy.orm import Session

from .base import BaseProductionML
from models.production import QualiteRecolte
from services.weather_service import WeatherService
from services.iot_service import IoTService

class QualitePredictor(BaseProductionML):
    """Service ML pour la prédiction de la qualité des récoltes"""
    
    def __init__(self, db: Session):
        super().__init__(db)
        self.weather_service = WeatherService(db)
        self.iot_service = IoTService(db)

    async def predict_qualite(
        self,
        parcelle_id: str,
        date_recolte: date
    ) -> Dict[str, Any]:
        """Prédit la qualité de la récolte"""
        # Récupération des données
        historique = await self._get_historique_qualite(parcelle_id)
        meteo = await self.weather_service.get_forecast(parcelle_id)
        iot_data = await self.iot_service.get_sensor_data(
            parcelle_id,
            date.today(),
            date_recolte
        )
        
        # Calcul des features
        features = self._calculate_quality_features(
            historique,
            meteo,
            iot_data
        )
        
        # Prédiction avec le modèle
        prediction = await self._predict_quality(features)
        
        return {
            "qualite_prevue": prediction["classe"],
            "probabilites": prediction["probabilites"],
            "facteurs_impact": await self._analyze_quality_factors(features)
        }

    def _calculate_quality_features(
        self,
        historique: List[Dict[str, Any]],
        meteo: List[Dict[str, Any]],
        iot_data: List[Dict[str, Any]]
    ) -> np.ndarray:
        """Calcule les features pour la prédiction de qualité"""
        features = []
        
        # Features historiques
        if historique:
            features.extend([
                len([h for h in historique if h["qualite"] == QualiteRecolte.A]) / len(historique),
                len([h for h in historique if h["qualite"] == QualiteRecolte.B]) / len(historique),
                len([h for h in historique if h["qualite"] == QualiteRecolte.C]) / len(historique)
            ])
        else:
            features.extend([0, 0, 0])
            
        # Features météo
        if meteo:
            features.extend([
                np.mean([m["temperature"] for m in meteo]),
                np.mean([m["humidite"] for m in meteo]),
                np.sum([m["precipitation"] for m in meteo])
            ])
        else:
            features.extend([0, 0, 0])
            
        # Features IoT
        if iot_data:
            features.extend([
                np.mean([d["valeur"] for d in iot_data]),
                np.std([d["valeur"] for d in iot_data]),
                len(iot_data)
            ])
        else:
            features.extend([0, 0, 0])
            
        return np.array(features)

    async def _predict_quality(
        self,
        features: np.ndarray
    ) -> Dict[str, Any]:
        """Prédit la qualité de la récolte"""
        # TODO: Implémenter le modèle de prédiction de qualité
        return {
            "classe": QualiteRecolte.A,
            "probabilites": {
                QualiteRecolte.A: 0.7,
                QualiteRecolte.B: 0.2,
                QualiteRecolte.C: 0.1
            }
        }

    async def _analyze_quality_factors(
        self,
        features: np.ndarray
    ) -> List[Dict[str, Any]]:
        """Analyse les facteurs influençant la qualité"""
        # TODO: Implémenter l'analyse des facteurs
        return [
            {
                "facteur": "Température",
                "impact": 0.4,
                "optimal": "25°C",
                "actuel": "28°C"
            },
            {
                "facteur": "Humidité",
                "impact": 0.3,
                "optimal": "70%",
                "actuel": "65%"
            }
        ]
